\name{hook_plot_htmlsubfig}
\alias{hook_plot_htmlsubfig}
\alias{hook_plot_beamer}
\title{Hooks for knitr}
\usage{
hook_plot_htmlsubfig(x, options)
hook_plot_beamer(x, options)
}
\arguments{
\item{x}{see \code{\link[knitr]{hook_pdfcrop}}}

\item{options}{see \code{\link[knitr]{hook_pdfcrop}}}
}
\description{
Hooks for including in knitr documents
}

